+++ src/App.module.css
@@ -1,0 +1,3 @@
+.mainSection {
+	margin: 16px;
+}
