+++ src/App.tsx
@@ -1,24 +1,17 @@
 import React from 'react';
 import Header from './components/Header';
-import RestaurantItem from './components/RestaurantItem';
+import RestaurantList from './components/RestaurantList';
 import SelectorSection from './components/SelectorSection';
-import mockData from './data/mockData.json';
-import type { Restaurant } from './components/RestaurantItem/type';
-
-interface mock {
-	restaurantList: Restaurant[];
-}
+import styles from './App.module.css';
 class App extends React.Component {
 	render() {
-		const { restaurantList } = mockData as mock;
-
 		return (
 			<main className='App'>
 				<Header />
-				<SelectorSection />
-				{restaurantList.map((restaurant) => {
-					return <RestaurantItem key={restaurant.id} restaurant={restaurant} />;
-				})}
+				<section className={styles.mainSection}>
+					<SelectorSection />
+					<RestaurantList />
+				</section>
 			</main>
 		);
 	}
