+++ src/components/RestaurantList/RestaurantList.module.css
@@ -1,0 +1,3 @@
+.restaurantList {
+	padding-top: 20px;
+}
